requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
aiosqlite==0.21.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
blockbuster==1.5.24
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==1.0.7
click==8.1.8
cloudpickle==3.1.1
coloredlogs==15.0.1
cryptography==44.0.2
dataclasses-json==0.6.7
Deprecated==1.2.18
distro==1.9.0
durationpy==0.9
fastapi==0.115.9
filelock==3.18.0
Flask==3.1.0
flatbuffers==25.2.10
forbiddenfruit==0.1.4
frozenlist==1.6.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.39.0
googleapis-common-protos==1.70.0
greenlet==3.2.1
grpcio==1.71.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jsonschema_rs==0.29.1
kubernetes==32.0.1
langchain==0.3.24
langchain-community==0.3.23
langchain-core==0.3.56
langchain-openai==0.3.14
langchain-text-splitters==0.3.8
langgraph==0.4.0
langgraph-api==0.1.22
langgraph-checkpoint==2.0.25
langgraph-checkpoint-sqlite==2.0.6
langgraph-cli==0.2.7
langgraph-prebuilt==0.1.8
langgraph-runtime-inmem==0.0.8
langgraph-sdk==0.1.64
langsmith==0.3.38
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.1.0
mpmath==1.3.0
multidict==6.4.3
mypy_extensions==1.1.0
narwhals==1.37.1
numpy==2.2.5
oauthlib==3.2.2
onnxruntime==1.21.1
openai==1.76.2
opentelemetry-api==1.32.1
opentelemetry-exporter-otlp-proto-common==1.32.1
opentelemetry-exporter-otlp-proto-grpc==1.32.1
opentelemetry-instrumentation==0.53b1
opentelemetry-instrumentation-asgi==0.53b1
opentelemetry-instrumentation-fastapi==0.53b1
opentelemetry-proto==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
opentelemetry-util-http==0.53b1
orjson==3.10.18
ormsgpack==1.9.1
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pillow==11.2.1
posthog==4.0.1
propcache==0.3.1
protobuf==5.29.4
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.24.0
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
sse-starlette==2.1.3
starlette==0.45.3
streamlit==1.45.0
structlog==25.3.0
sympy==1.14.0
tavily-python==0.7.0
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typer==0.15.3
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchdog==6.0.0
watchfiles==1.0.5
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wikipedia==1.4.0
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
zstandard==0.23.0
